+++ src/components/publicProfile.js
@@ -5,13 +5,13 @@ import { navigate } from "gatsby"
 import LearningPathCard from "./User/learningPathCard"
 import { Row, Col, Icon } from "antd"
 
-import Header from "./Layout/sideDrawer"
-import Footer from "./Layout/footer"
-import Loader from "./Layout/loader"
+import Header from "../Stateless/Layout/sideDrawer"
+import Footer from "../Stateless/Layout/footer"
+import Loader from "../Stateless/Layout/loader"
 
-import userStyles from "../styles/user.module.css"
-import styles from "../styles/layout.module.css"
-import pathStyles from "../styles/result.module.css"
+import Userstyles from "../styles/user.module.css"
+import Layoutstyles from "../styles/layout.module.css"
+import Resultstyles from "../styles/result.module.css"
 
 export class publicProfile extends Component {
   state = {
@@ -34,7 +34,7 @@ export class publicProfile extends Component {
           if (loading)
             return (
               <div>
-                <div className={styles.content}>
+                <div className={Layoutstyles.content}>
                   <Header show={false} />
                   <Loader />
                 </div>
@@ -44,7 +44,7 @@ export class publicProfile extends Component {
           if (error)
             return (
               <div>
-                <div className={styles.content}>
+                <div className={Layoutstyles.content}>
                   <Header show={false} />
                   <h1>Error Loading User</h1>
                 </div>
@@ -55,38 +55,38 @@ export class publicProfile extends Component {
             <div>
               {data ? (
                 <div>
-                  <div className={userStyles.content}>
+                  <div className={Userstyles.content}>
                     <Header show={false} />
-                    <div className={userStyles.container}>
+                    <div className={Userstyles.container}>
                       <div
                         className={
-                          userStyles.profileContainer +
+                          Userstyles.profileContainer +
                           " " +
-                          userStyles.desktopProfile
+                          Userstyles.desktopProfile
                         }
                       >
-                        <div className={userStyles.profileInfo}>
+                        <div className={Userstyles.profileInfo}>
                           <h1>
                             {data.first_name} {data.last_name}
                           </h1>
                           <h2>{data.username}</h2>
                           <h3>{data.bio}</h3>
                         </div>
-                        <div className={userStyles.profileImgContainer}>
+                        <div className={Userstyles.profileImgContainer}>
                           <img
-                            className={userStyles.profileImg}
+                            className={Userstyles.profileImg}
                             src={data.profile_pic}
                             alt=""
                           />
-                          <div className={userStyles.social}>
-                            <Row className={userStyles.social_row}>
+                          <div className={Userstyles.social}>
+                            <Row className={Userstyles.social_row}>
                               {data.github !== "" &&
                               data.github !== null &&
                               data.github !== "https://github.com/" ? (
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.github)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="github"
                                     theme="filled"
                                   />
@@ -100,7 +100,7 @@ export class publicProfile extends Component {
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.linkedin)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="linkedin"
                                     theme="filled"
                                   />
@@ -114,7 +114,7 @@ export class publicProfile extends Component {
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.twitter)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="twitter"
                                   />
                                 </Col>
@@ -127,7 +127,7 @@ export class publicProfile extends Component {
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.facebook)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="facebook"
                                     theme="filled"
                                   />
@@ -138,18 +138,18 @@ export class publicProfile extends Component {
                             </Row>
                           </div>
                         </div>
-                        <div className={userStyles.profileDetail}>
-                          <div className={userStyles.profileAbout}>
+                        <div className={Userstyles.profileDetail}>
+                          <div className={Userstyles.profileAbout}>
                             <h1>About</h1>
-                            <p className={userStyles.bioContent}>
+                            <p className={Userstyles.bioContent}>
                               {data.about}
                             </p>
                             {data.skills ? (
-                              <div className={userStyles.skills}>
+                              <div className={Userstyles.skills}>
                                 <h1>Skills</h1>
                                 {data.skills.split(",").map((skill, i) => {
                                   return (
-                                    <div key={i} className={userStyles.skill}>
+                                    <div key={i} className={Userstyles.skill}>
                                       {skill}
                                     </div>
                                   )
@@ -159,37 +159,37 @@ export class publicProfile extends Component {
                               <div></div>
                             )}
                           </div>
-                          <Row className={userStyles.userInfo}>
+                          <Row className={Userstyles.userInfo}>
                             <Col span={8}>
-                              <div className={userStyles.count}>
+                              <div className={Userstyles.count}>
                                 {data.learning_paths_aggregate.aggregate.count}
-                                <span className={userStyles.text}>
+                                <span className={Userstyles.text}>
                                   Learning Paths
                                 </span>
                               </div>
                             </Col>
                             <Col span={8}>
-                              <div className={userStyles.count}>
+                              <div className={Userstyles.count}>
                                 {data.followers_aggregate.aggregate.count}
-                                <span className={userStyles.text}>
+                                <span className={Userstyles.text}>
                                   Followers
                                 </span>
                               </div>
                             </Col>
                             <Col span={8}>
-                              <div className={userStyles.count}>
+                              <div className={Userstyles.count}>
                                 {data.following_aggregate.aggregate.count}
-                                <span className={userStyles.text}>
+                                <span className={Userstyles.text}>
                                   Following
                                 </span>
                               </div>
                             </Col>
                           </Row>
                         </div>
-                        <div className={userStyles.profileLearningPath}>
+                        <div className={Userstyles.profileLearningPath}>
                           <h1>Learning Paths ({data.learning_paths.length})</h1>
                         </div>
-                        <div className={userStyles.profilePath}>
+                        <div className={Userstyles.profilePath}>
                           {data.learning_paths.map((data, index) => {
                             return (
                               <LearningPathCard
@@ -202,14 +202,14 @@ export class publicProfile extends Component {
                         </div>
                       </div>
 
-                      <div className={userStyles.mobileProfile}>
-                        <div className={userStyles.profileContainer}>
+                      <div className={Userstyles.mobileProfile}>
+                        <div className={Userstyles.profileContainer}>
                           <img
-                            className={userStyles.avatar}
+                            className={Userstyles.avatar}
                             src={data.profile_pic}
                             alt=""
                           />
-                          <div className={userStyles.userDetails}>
+                          <div className={Userstyles.userDetails}>
                             <h2>
                               {data.first_name} {data.last_name}
                             </h2>
@@ -217,42 +217,42 @@ export class publicProfile extends Component {
                             <h4>{data.bio}</h4>
                           </div>
 
-                          <div className={userStyles.userInfo}>
+                          <div className={Userstyles.userInfo}>
                             <Row>
                               <Col span={8}>
-                                <div className={userStyles.count}>
+                                <div className={Userstyles.count}>
                                   {data.followers_aggregate.aggregate.count}
                                 </div>
-                                <div className={userStyles.text}>Followers</div>
+                                <div className={Userstyles.text}>Followers</div>
                               </Col>
                               <Col span={8}>
-                                <div className={userStyles.count}>
+                                <div className={Userstyles.count}>
                                   {data.following_aggregate.aggregate.count}
                                 </div>
-                                <div className={userStyles.text}>Following</div>
+                                <div className={Userstyles.text}>Following</div>
                               </Col>
                               <Col span={8}>
-                                <div className={userStyles.count}>
+                                <div className={Userstyles.count}>
                                   {
                                     data.learning_paths_aggregate.aggregate
                                       .count
                                   }
                                 </div>
-                                <div className={userStyles.text}>
+                                <div className={Userstyles.text}>
                                   Learning Paths
                                 </div>
                               </Col>
                             </Row>
                           </div>
 
-                          <div className={userStyles.bio}>{data.about}</div>
+                          <div className={Userstyles.bio}>{data.about}</div>
 
                           {data.skills !== null ? (
-                            <div className={userStyles.skills}>
+                            <div className={Userstyles.skills}>
                               <h2>Skills</h2>
                               {data.skills.split(",").map((skill, i) => {
                                 return (
-                                  <div key={i} className={userStyles.skill}>
+                                  <div key={i} className={Userstyles.skill}>
                                     {skill}
                                   </div>
                                 )
@@ -262,15 +262,15 @@ export class publicProfile extends Component {
                             <div></div>
                           )}
 
-                          <div className={userStyles.social}>
-                            <Row className={userStyles.social_row}>
+                          <div className={Userstyles.social}>
+                            <Row className={Userstyles.social_row}>
                               {data.github !== "" &&
                               data.github !== null &&
                               data.github !== "https://github.com/" ? (
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.github)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="github"
                                     theme="filled"
                                   />
@@ -284,7 +284,7 @@ export class publicProfile extends Component {
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.linkedin)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="linkedin"
                                     theme="filled"
                                   />
@@ -298,7 +298,7 @@ export class publicProfile extends Component {
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.twitter)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="twitter"
                                   />
                                 </Col>
@@ -311,7 +311,7 @@ export class publicProfile extends Component {
                                 <Col span={6}>
                                   <Icon
                                     onClick={() => window.open(data.facebook)}
-                                    className={userStyles.icon}
+                                    className={Userstyles.icon}
                                     type="facebook"
                                     theme="filled"
                                   />
@@ -323,12 +323,12 @@ export class publicProfile extends Component {
                           </div>
                         </div>
                         <Row>
-                          <div className={userStyles.mobileProfileLearningPath}>
+                          <div className={Userstyles.mobileProfileLearningPath}>
                             <h1>
                               Learning Paths ({data.learning_paths.length})
                             </h1>
                           </div>
-                          <div className={pathStyles.resultPath}>
+                          <div className={Resultstyles.resultPath}>
                             {data.learning_paths.map((data, index) => {
                               return (
                                 <LearningPathCard
@@ -348,12 +348,14 @@ export class publicProfile extends Component {
                 </div>
               ) : (
                 <div>
-                  <div className={styles.content}>
+                  <div className={Layoutstyles.content}>
                     <Header show={false} />
                     <div>
-                      <h1 className={styles.public_not_found}>
+                      <h1 className={Layoutstyles.public_not_found}>
                         No profile found with username{" "}
-                        <span className={styles.public_not_found_username}>
+                        <span
+                          className={Layoutstyles.public_not_found_username}
+                        >
                           {this.props.username}
                         </span>
                       </h1>
@@ -365,7 +367,7 @@ export class publicProfile extends Component {
                         }}
                       >
                         <div
-                          className={styles.public_not_found_go_home}
+                          className={Layoutstyles.public_not_found_go_home}
                           onClick={() => navigate("/")}
                         >
                           <Icon type="home" style={{ marginRight: "10px" }} />{" "}
