+++ src/pages/add.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react"
 
-import Layout from "../components/Layout/layout"
-import AddPath from "../components/Create/addPath"
+import Layout from "../Statefull/Layout/layout"
+import AddPath from "../Statefull/Create/addPath"
 
 export class add extends Component {
   render() {
