+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react"
 
-import Layout from "../components/Layout/layout"
-import Landing from "../components/Landing"
+import Layout from "../Statefull/Layout/layout"
+import Landing from "../Stateless/Landing/index"
 
 export class index extends Component {
   render() {
