+++ src/pages/profile.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react"
-import Layout from "../components/Layout/layout"
-import User from "../components/User"
+import Layout from "../Statefull/Layout/layout"
+import User from "../Stateless/User/index"
 
 export class profile extends Component {
   render() {
