+++ src/utils/typography.js
@@ -1,4 +1,6 @@
 import Typography from "typography"
 import grandViewTheme from "typography-theme-grand-view"
+
 const typography = new Typography(grandViewTheme)
+
 export default typography
